+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Ability to Exclude Folders from RAG (https://github.com/nbonamy/witsy/issues/515)
 - Stop agent execution buttons
+- Ollama image generation support (https://github.com/nbonamy/witsy/issues/521)
 
 ### Changed
 - New agent view with step-by-step execution